requirements.txt
openai==1.13.3
python-dotenv==1.0.1
requests==2.28.2